More compact encoding for VM instructions
The Split instruction always has the immediately following instruction
as one of its two branch targets. The only question is which of the
branch targets should be given higher priority, and that can be encoded
in one bit. Because most of the enum payloads are 2-byte aligned and the
tag only needs one byte, there's a spare padding byte to put this flag
in, shrinking the enum from 6 bytes to 4.

This also simplifies compiling a regex because the `greedy` flag on
repetition operators determines which target to prefer.