Fix non-greedy repetition
I got too clever when trying to compile multiple patterns into a single
bytecode program.

When a thread matches, we need to discard all lower-priority threads,
but only for threads which are from the same pattern. Mixing all the
threads from separate patterns together made that difficult, though not
impossible.

But instead I've chosen to compile one pattern at a time, and provide an
function to execute many compiled programs on the same input. That
simplifies a lot of the implementation and lets me provide a better API
too.