Transform Pike VM programs to DFAs
This is essentially just the standard power-set construction for NFA to
DFA transformation, except that zero-width assertions and capture groups
introduce some trickiness.

Edges in this DFA carry mini-programs to implement the side-effects of
the Pike VM instructions that the DFA is simulating. Those side effects
can typically be simplified by treating the DFA as a control-flow graph
and using standard compiler optimizations. See papers from the re2c
project for a bunch of optimizations that are helpful in this specific
setting.

Zero-width assertions aren't quite right yet; the set of active
instructions needs to be qualified by any constraints those instructions
were visited with.

This passes all of the current tests, which of course means the tests
aren't yet comprehensive enough.