Fix optional failed assertions in Pike VM
When simulating an NFA using a Pike-style VM, zero-width assertions like
`^` must be checked immediately, not deferred until the next Range or
Match instruction. Otherwise, if there are two paths to an instruction,
where the higher-priority path passes through an assertion that fails,
we'd conclude that the instruction is unreachable, even though it may be
reachable through the lower-priority path.

This fix actually made the implementation quite a bit simpler! In order
to check the assertion in Threads::add, I needed to pass the previous
and next byte of input into that function. But that meant I could also
check the Range and Match cases there, rather than handling some
VM instructions there and others in `Program::exec_many`.