Fix optional failed assertions for DFA
Like the previous commit's fix for this issue in my Pike VM interpreter,
my DFA builder was overly aggressive about merging states which were
actually distinguishable on some inputs. It just wasn't as obvious
because I'd started thinking about this issue while writing the DFA
builder so I got some of the cases right.

This is now _less_ aggressive than it could be, so there are ways to
generate some non-minimal DFAs. But there already were ways to do that,
and a DFA minimization pass would clean these new cases up too. Still,
with a bit of thought someone could probably get the builder to generate
smaller DFAs in the first place.