Fix optional failed assertions for DFA, better
Fuzz testing found an example, minimized to `\B\z|\z`, where the
constraints on both the previous and the next input matter when deciding
whether it's safe to discard states during DFA construction.

I suspected there might be cases like that but hadn't come up with a
specific one yet, so it was nice when the fuzzer found it for me.