Copy all nullable subexpressions in nullable replay
PCRE has weird behavior for repetitions of nullable subexpressions. It
matches the subexpression once more after the repetition can't consume
any more of the input, so any groups that capture within that
subexpression when matched on the empty string are set to point after
the end of the match.

In this commit I return to my earlier strategy of recursively replaying
the subexpression, but filtering it to only those parts which can match
on an empty string. It turns out not to be enough to replay just the
Save instructions for capture groups: anchors and other assertions can
lead to different capture groups being selected, depending on
surrounding context. So now I just replay everything that doesn't add
any Range instructions.

This is grossly wasteful when it happens, but I'm tired of fighting with
a case that I don't think is useful. Doing it this way keeps the hacks
from spreading across the rest of the compiler. After DFA conversion,
the redundant edges are cleaned up anyway, and someday, CFG optimization
should clean up the redundant register writes.